+++ legacy/l/lushootseed/source/help/lushootseed.php
@@ -1,0 +1,11 @@
+<?php 
+  $pagename = 'Lushootseed Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+
+<p>
+    Documentation for this keyboard is available here: <a href="lushootseed unicode keyboard help.pdf">lushootseed unicode keyboard help.pdf</a>
+</p>