+++ legacy/m/moore/source/help/moore.php
@@ -1,0 +1,69 @@
+﻿<?php 
+  $pagename = 'Mooré Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+<h1>Moor gʋndga</h1>
+<p>Gʋndg kãngã naana moor gʋls-biis la gom-biis yĩnga.</p>
+<p>B yag-a-la fãrendã gʋndg zugu.</p>
+<p>Moorã gʋndg mamseg lagma teed a taab n be kaane. Rẽ, bɩ-y pak mamsegã n bãng gʋls-biisã zĩisi.</p>
+<p>Gʋndgã naagda ne <i>Doulos SIL</i> gʋls-wʋng wala <i>Charis SIL</i> n tʋmdẽ.</p>
+<p><i>Doulos SIL</i> wala <i>Charis SIL</i> gʋls-wʋmdɩ wã lagma ne teed a taabã n be ka. Y tõe n paama yaolem gʋls-wʋmdɩ tilgẽ ka :
+<a href='http://software.sil.org/'>http://software.sil.org/</a>.</p>
+
+
+<h1>Clavier mooré</h1>
+
+<p>Ce clavier est destiné à la saisie et à l’affichage de textes/documents en langue mooré. Il est construit 
+à partir du clavier AZERTY français. Une infographie incluse dans ce pack permet de repérer les différents 
+emplacements des caractères spécifiques au mooré.</p>
+
+<p>Ce clavier doit être utilisé conjointement avec une police de caractères comme <i>Doulos SIL</i> ou <i>Charis SIL</i>.</p>
+
+<p>Les polices <i>Doulos SIL</i> et <i>Charis SIL</i> sont incluses dans ce pack. Vous pouvez effectuer des mises à jour 
+de ces polices de caractères depuis le site <a href='http://software.sil.org/'>http://software.sil.org/</a>.</p>
+
+<h2>Notes</h2>
+<ul>
+  <li>La touche- commande des tons est localisée en haut à l’extrême gauche du clavier :
+    <ul>
+      <li><span class='g'>◌̀</span>: ton bas		² </li>
+      <li><span class='g'>◌́</span>: ton haut		Shift + ² </li>
+      <li><span class='g'>◌̄</span>: ton moyen		AltGr + ² </li>
+    </ul>
+  </li>
+  <li>
+    Les caractères spéciaux pour linguistes (<span class='g'>ŋ ʔ ɖ ɗ ɦ ɟ ɠ ɲ ɳ</span>) sont obtenus par modification du caractère
+    de base par AltGr et Shift+AltGr. Il faut noter que <span class='g'>ŋ</span> est obtenu par modification de G (AltGr+G,) et non par AltGr+N.
+  </li>
+  <li>
+    Chaque caractère est stocké en tant que forme composée et les marques tonales des diacritiques se combinant ave les caratères pleins.
+  </li>
+</ul>
+
+  <h1>Mooré Keyboard Layout</h1>
+  
+  <p>This keyboard is designed for input of the Mooré language.  It is based on the French keyboard as used in Burkina
+  Faso.  A visual keyboard is included in this package; please refer to it for placement of specific characters.</p>
+  
+  <p>This keyboard should be used with a font such as <i>Charis SIL</i> or <i>Doulos SIL</i>.  <i>Charis SIL</i> is included in this package, and
+  updated versions of all these fonts can be downloaded from <a href='http://software.sil.org/'>http://software.sil.org/</a>.</p>
+    
+  <h2>Notes</h2>
+  
+  <ul>
+    <li>Tone marks are located on the top-left key of the keyboard:
+      <ul><li><span class='g'>◌́</span>: High tone - press Shift + ²</li>
+      <li><span class='g'>◌̀</span>: Low tone - press ²</li>
+      <li><span class='g'>◌̄</span>: Mid tone - press AltGr + ²</li></ul>
+    </li>
+    
+    <li>Additional linguistic characters (<span class='g'>ŋ ʔ ɖ ɗ ɦ ɟ ɠ  ɲ ɳ</span>) are located on their associated base keys, 
+      modified with AltGr and Shift+AltGr.  Note that <span class='g'>ŋ</span> is located on AltGr+G, not AltGr+N.</li>
+      
+    <li>When keys have been modified from the original Burkina Faso layout, where possible, they have been moved to the AltGr state of the same key.</li>
+    
+    <li>All characters are stored in composed form, and tone marks are stored as combining diacritics.</li>
+  </ul>