+++ legacy/u/urdu_unicode/source/help/urdu_unicode.php
@@ -1,0 +1,11 @@
+<?php 
+  $pagename = 'Urdu Utrans-to-Unicode Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+
+<p>
+    Documentation for this keyboard is available here: <a href="urdu_unicode.pdf">urdu_unicode.pdf</a>
+</p>