+++ release/basic/basic_kbdest/source/help/basic_kbdest.php
@@ -1,0 +1,27 @@
+<?php 
+  $pagename = 'Estonian Basic Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+<p>
+    This keyboard layout is designed for Estonian.  It includes an on screen keyboard which can be viewed by clicking on the Keyman icon and selecting the On Screen Keyboard menu item. The keyboard layout follows the Windows 10 Estonian layout.
+</p>
+<p>This keyboard makes use of "deadkeys" which are listed below:</p>
+
+<p>´c will produce ć (available on cCeEnNoOsSzZ).</p>
+<p>`a will produce à (available on aAeEoOuU).</p>
+<p>^a will produce â (available on aAeEgGiIuU).</p>
+<p>ˇc will produce č (available on cCsSzZ).</p>
+<p>~o will produce õ (available on oO).</p>
+
+<h1>Keyboard Layout</h1>
+
+<h2>Desktop Keyboard Layout</h2>
+<div id='osk' data-states='default shift rightalt rightalt-shift'>
+</div>
+
+<h2>Mobile/Tablet Keyboard Layout</h2>
+
+<div id='osk-tablet' data-states='default shift rightalt'>
+</div>