+++ release/basic/basic_kbdyba/source/help/basic_kbdyba.php
@@ -13,9 +13,14 @@ This keyboard is designed for the Yoruba language of Nigeria.
 </p>
 <p>If square boxes are displayed instead of characters when using this keyboard (and in the keyboard layouts below), please read our <a href="/troubleshooting/#boxes">troubleshooting guide</a>.
 </p>
-<p>Acute accents can be typed by using the deadkey "'" followed by a vowel or "n". Thus, typing "'a" will produce á. To access the "'" character, type a space after it.</p>
-<p>Grave accents can be typed by using the deadkey "[" followed by a vowel or "n". Thus, typing "[a" will produce à. To access the '[' character, type a space after it.</p>
-<p>Dot below combining marks can be typed by using the deadkey "\" followed by an "e" or "o" vowel or "s". Thus, typing "\e" will produce ẹ. To access the '\' character, type a space after it.</p>
+
+<p>This keyboard makes use of "deadkeys" which are listed below:</p>
+
+<p>[a will produce à (available on aAeEiImMnNoOuU).</p>
+<p>\e will produce ẹ (available on eEoOsS).</p>
+<p>'a will produce á (available on aAeEiImMnNoOuU).</p>
+
+<p>To access the "deadkey" itself, type a space after it.</p>
 
 <br/>