+++ release/g/gondi_gunjala/HISTORY.md
@@ -1,7 +1,12 @@
 Gondi (Gunjala) Keyboard Change History
 =======================================
 
+1.1 (17 Feb 2020)
+-----------------
+* Added released version of the font to the keyboard package
+* No change to the keyboard itself
+
 1.0 (24 Aug 2019)
 -----------------
-* Initial design for harware keyboard based on the WinScript phonetic layout
+* Initial design for hardware keyboard based on the WinScript phonetic layout
 * MP: added a unique touch interface optimized for the Gondi language